refactor: Extract mixin for executable search paths
