+++ src/python/pants/core/util_rules/search_paths.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 import logging
+import os
 from dataclasses import dataclass
 from pathlib import Path
 from typing import Iterable
@@ -10,10 +11,13 @@ from typing import Iterable
 from pants.base.build_environment import get_buildroot
 from pants.core.util_rules.environments import EnvironmentTarget, LocalEnvironmentTarget
 from pants.engine.collection import DeduplicatedCollection
+from pants.engine.env_vars import EnvironmentVars
 from pants.engine.rules import Rule, _uncacheable_rule, collect_rules, rule
+from pants.option.option_types import StrListOption
 from pants.util.logging import LogLevel
-from pants.util.ordered_set import FrozenOrderedSet
-from pants.util.strutil import softwrap
+from pants.util.memo import memoized_property
+from pants.util.ordered_set import FrozenOrderedSet, OrderedSet
+from pants.util.strutil import help_text, softwrap
 
 _logger = logging.getLogger(__name__)
 
@@ -140,6 +144,36 @@ async def validate_search_paths(request: ValidateSearchPathsRequest) -> Validate
         )
 
     return ValidatedSearchPaths(search_paths)
+
+
+class ExecutableSearchPathsOptionMixin:
+    executable_search_paths_help: str
+    _options_env: EnvironmentVars
+
+    executable_search_paths = StrListOption(
+        default=["<PATH>"],
+        help=lambda cls: help_text(
+            f"""
+            {cls.executable_search_paths_help}
+            The special string `"<PATH>"` will expand to the contents of the PATH env var.
+            """
+        ),
+        advanced=True,
+        metavar="<binary-paths>",
+    )
+
+    @memoized_property
+    def executable_search_path(self) -> tuple[str, ...]:
+        def iter_path_entries():
+            for entry in self.executable_search_paths:
+                if entry == "<PATH>":
+                    path = self._options_env.get("PATH")
+                    if path:
+                        yield from path.split(os.pathsep)
+                else:
+                    yield entry
+
+        return tuple(OrderedSet(iter_path_entries()))
 
 
 def rules() -> Iterable[Rule]:
